create bootstrap target in Makefile
Also, add commands in bootstrap.sh that treat files in etc/, which are
now available in the tarball created by `make bootstrap`. The intended
use-case is like

    make bootstrap
    scp bootstrapit.tgz USER@HOST:/home/USER/

and on HOST, untar and run bootstrap.sh. The user on HOST does not
need to be root, but beware of auth timeouts between `sudo` calls in
bootstrap.sh.