update nginx configuration assuming nv.rerobots
This is the path to the current (live) certificate as provided by
Let's Encrypt (https://letsencrypt.org/).

The only deployment at the time of this commit is to
https://nv.rerobots.net
This can be easily changed in nginx.conf if you have a different
domain name.

also, do not use OCSP Stapling yet