consider sim job as passing if call to set_pwm
If a call to PCA9685.set_pwm() is detected, then end simulation job
and mark it as `OK`. If more than one instance of PCA9685 exists, then
the set_pwm() method of at least one must be called.

For example, this occurs during the variations of

    python manage.py drive

in the Donkey car guide at https://docs.donkeycar.com/guide/get_driving/